+++ crates/gen/src/lib.rs
@@ -1,4 +1,4 @@
-mod case;
+mod to_snake;
 mod winmd;
 mod type_limits;
 mod type_namespaces;
@@ -12,7 +12,8 @@ pub use type_limits::{NamespaceTypes, TypeLimit, TypeLimits};
 pub use type_reader::TypeReader;
 pub use type_tree::TypeTree;
 pub use types::{Type,Class,Interface,Delegate,Struct, Enum, TypeName};
-use winmd::*;
+pub use winmd::*;
+use to_snake::to_snake;
 
 fn format_ident(name: &str) -> squote::Ident {
     if name == "Self" {
