+++ crates/gen/src/to_snake.rs
