+++ crates/gen/src/types/type_name.rs
@@ -1,5 +1,4 @@
 use crate::Blob;
-use crate::*;
 use crate::winmd::*;
 use crate::types::*;
 use crate::*;
