+++ crates/gen/src/winmd/codes.rs
@@ -1,4 +1,3 @@
-use crate::*;
 use crate::Row;
 use crate::winmd::*;
 use crate::TypeReader;
